Removed Jekyll theme.
Delete the one line. Not sure if there's a better way to start GitHub Pages with Markdown (instead of regular HTML), but I chose a theme first so it read content from the "README.md" file (instead of an index.html which I hadn't created yet), and then deleted the line for the theme in this "_config.yml".